Second pass on twentyten/functions.php inline documentation. see #12695. Also:
rename twentyten_init() to twentyten_setup(), to reflect the hook it is run on (init would be too late, it must be after_setup_theme). Remove unnecessary twentyten_get_page_number() and transfer functionality to twentyten_the_page_number(). Remove the function_exists() wrappers from functions that are simply tied to a hook, as a better practice would be to remove the hook instead of plugging the function. For architecture changes, see #12748.

git-svn-id: https://develop.svn.wordpress.org/trunk@13886 602fd350-edb4-49c9-b593-d223f7449a82