Improve inline documentation in `wp_ajax_destroy_sessions()`.
If the token is set (`$keep` is a string), this means the user is viewing their own profile-editing screen and destroying their own sessions (except the current one). If it isn't set (`$keep` is null), the user is editing another user's profile and destroying all of their sessions with no exceptions.

See #30264, #30469


git-svn-id: https://develop.svn.wordpress.org/trunk@30595 602fd350-edb4-49c9-b593-d223f7449a82