Multisite: Restore display of the Toolbar on wp-activate.php while logged-in.
Just as with wp-signup.php, display of the Toolbar on wp-activate.php was broken in [23512] due to a hook change for the `_wp_admin_bar_init()` callback. wp-signup.php was fixed in [35423] for #34418.

See [35423] for more in-depth background and reasoning behind restoring display of the Toolbar on wp-activate.php and wp-signup.php.

Fixes #34496. See #34418.


git-svn-id: https://develop.svn.wordpress.org/trunk@35449 602fd350-edb4-49c9-b593-d223f7449a82