Remove PHP 7 from allowed failures list
PHP 7 has released it's 7th (and potentially final) Release Candidate today. While it is still not recommended for use in production, it is at a point that a test failure with it is something that needs to be triaged immediately. Congratulations to the PHP internals team on the upcoming final release of PHP 7.

Fixes #33410



git-svn-id: https://develop.svn.wordpress.org/trunk@35633 602fd350-edb4-49c9-b593-d223f7449a82